pyproject.toml
[project]
name = "local-vendor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "jwt>=1.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
